organize tests into proper directories and rename tests
