+++ version0/tests/mocks/mock_file_logger.py
@@ -1,6 +1,6 @@
 from collections import deque
 
-from logger.mock_log_result import MockLogResult
+from tests.mocks.mock_log_result import MockLogResult
 
 
 class MockFileLogger:
