+++ version0/tests/mocks/mock_log_result.py
